refactor(llment): move prompt assets into command
